% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose a MultiAssayExperiment object and write the components to disk}
\usage{
decompose(object, prepend = "myMAE_", directory = ".")
}
\arguments{
\item{object}{An object of class MultiAssayExperiment}

\item{prepend}{A character string to pre-pend to the filename}

\item{directory}{The directory where files will be written to}
}
\value{
A character vector of the filenames created
}
\description{
Decompose a MultiAssayExperiment object and write the components to disk
}
\examples{
example("MultiAssayExperiment")
res <- decompose(myMultiAssayExperiment)
res
}

