% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCuratedTCGAData.R
\name{getCuratedTCGAData}
\alias{getCuratedTCGAData}
\title{Download TCGA data types from ExperimentHub}
\usage{
getCuratedTCGAData(cohort = "ACC", RNASeqGene = TRUE,
  RNASeq2GeneNorm = TRUE, miRNASeqGene = TRUE, CNASNP = TRUE,
  CNVSNP = TRUE, CNAseq = TRUE, Methylation = TRUE, RPPAArray = TRUE,
  Mutations = TRUE, gistica = TRUE, gistict = TRUE)
}
\arguments{
\item{cohort}{TCGA cohort disease code (e.g., COAD)}
}
\description{
Download TCGA data types from ExperimentHub
}

