% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae2rd.R
\name{mae2rd}
\alias{mae2rd}
\title{Write a .Rd man page for a MultiAssayExperiment object}
\usage{
mae2rd(object, filename, objname, title = objname, aliases = objname,
  descriptions = NULL)
}
\arguments{
\item{object}{An object of class MultiAssayExperiment}

\item{filename}{Full path of the filename of the .Rd man page to write}

\item{objname}{Name of the object being documented}

\item{aliases}{A list of aliases}

\item{descriptions}{A list of extra lines to be written to the Description}
}
\value{
The filename is returned after successfully writing the man file
}
\description{
Write a .Rd man page for a MultiAssayExperiment object
}
\examples{
example("MultiAssayExperiment")
res <- mae2rd(myMultiAssayExperiment, filename = file.path(tempdir(), "MyMAE.Rd"), objname="MyMAE")
res
}

