% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyTCGA.R
\name{simplifyTCGAData}
\alias{simplifyTCGAData}
\title{Simplify curatedTCGAData objects by replacing RaggedExperiment objects}
\usage{
simplifyTCGAData(obj, removeRaggedExperiments = TRUE)
}
\arguments{
\item{obj}{A MultiAssayExperiment object obtained from curatedTCGAData}

\item{removeRaggedExperiments}{if TRUE (default), remove RaggedExperiment objects
from the returned MultiAssayExperiment}
}
\value{
A MultiAssayExperiment object with RaggedExperiments converted to RangedSummarizedExperiment
with rows corresponding to gene symbole.

"Mutations" objects become a genes x patients RangedSummarizedExperiment containing 1 if
there is a non-silent mutation somewhere in the gene, and 0 otherwise.
"CNA" and "CNV" segmented copy number are reduced using a weighted mean in the rare
cases of overlapping (non-disjoint) copy number regions.
}
\description{
Simplify curatedTCGAData objects by replacing RaggedExperiment objects
}
\details{
Relies on TxDb.Hsapiens.UCSC.hg19.knownGene and org.Hs.eg.db to map to hg19 NCBI build.
}
\examples{
accmae <- curatedTCGAData("ACC", c("CNASNP", "Mutation"), dry.run = FALSE)
simplifyTCGAData(accmae)
}
