% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyTCGA.R
\name{symbolsToRanges}
\alias{symbolsToRanges}
\title{Convert SummarizedExperiment elements with gene symbols to RangedSummarizedExperiment}
\usage{
symbolsToRanges(obj, removeOriginals = TRUE)
}
\arguments{
\item{obj}{A MultiAssayExperiment object obtained from curatedTCGAData}

\item{removeOriginals}{If TRUE (default), remove the SummarizedExperiment objects
that have been converted to RangedSummarizedExperiment}
}
\value{
a MultiAssayExperiment where any of the original SummarizedExperiment containing
gene symbols as rownames have been replaced or supplemented by a RangedSummarizedExperiment
for miR that could be mapped to GRanges, and another SummarizedExperiment for miR that
could not be mapped to GRanges.
}
\description{
Convert SummarizedExperiment elements with gene symbols to RangedSummarizedExperiment
}
\details{
Any SummarizedExperiment elements with gene symbols as rownames will have ranges added.
Symbols where ranges can't be found are put in a new SummarizedExperiment.
}
\examples{
symbolsToRanges(miniACC)
}
\seealso{
mirToRanges
}
