% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTCGAData.R
\name{curatedTCGAData}
\alias{curatedTCGAData}
\title{Create a MultiAssayExperiment from specific assays and cohorts}
\usage{
curatedTCGAData(diseaseCode = "*", assays = "*", dry.run = TRUE)
}
\arguments{
\item{assays}{a character vector containing the name(s) of TCGA assays}

\item{cohort}{a character vector containing the name(s) of TCGA cohorts}
}
\value{
a MultiAssayExperiment of the specified assays and cohorts
}
\description{
Create a MultiAssayExperiment from specific assays and cohorts
}
\examples{
get_curatedTCGAData(cohort = "ACC", assays = c("RNASeqGene",
"RNASeq2GeneNorm"))
}
