% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTCGAData.R
\name{curatedTCGAData}
\alias{curatedTCGAData}
\title{Create a MultiAssayExperiment from specific assays and cohorts}
\usage{
curatedTCGAData(diseaseCode = "*", assays = "*", dry.run = TRUE)
}
\arguments{
\item{diseaseCode}{a character vector containing the name(s) of TCGA cohorts}

\item{assays}{a character vector containing the name(s) of TCGA assays}

\item{dry.run}{logical (default TRUE) whether to return the dataset names
before actual download}
}
\value{
a \linkS4class{MultiAssayExperiment} of the specified assays and
cancer codes
}
\description{
Create a MultiAssayExperiment from specific assays and cohorts
}
\details{
This function will check against available resources in
ExperimentHub. Currently, only the latest runDate ("2016-01-28") is
supported. Use the \code{dry.run = FALSE} to download remote datasets and
build an integrative \linkS4class{MultiAssayExperiment} object.
}
\examples{
curatedTCGAData(diseaseCode = c("GBM", "ACC"), assays = "CNASNP")

}
