% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCuratedTCGAData.R
\name{getCuratedTCGAData}
\alias{getCuratedTCGAData}
\title{Create a MultiAssayExperiment from specific assays and cohorts}
\usage{
getCuratedTCGAData(cohort = "ACC", assays = c("RNASeqGene",
  "RNASeq2GeneNorm", "miRNASeqGene", "CNASNP", "CNVSNP", "CNAseq",
  "Methylation", "RPPAArray", "Mutations", "gistica", "gistict"))
}
\arguments{
\item{cohort}{a character vector containing the name(s) of TCGA cohorts}

\item{assays}{a character vector containing the name(s) of TCGA assays}
}
\value{
a MultiAssayExperiment of the specified assays and cohorts
}
\description{
Create a MultiAssayExperiment from specific assays and cohorts
}
\examples{
get_curatedTCGAData(cohort = "ACC", assays = c("RNASeqGene",
"RNASeq2GeneNorm"))
}

